# Socify-Companion/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SOCIFY API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List campaigns
      parameters:
        - name: workspaceId
          in: query
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{campaignId}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get campaign
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update campaign
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCampaignBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete campaign
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── LISTENING ──────────────────────────────────────────────────────────────
...
  /automations:
    get:
      operationId: listAutomations
      tags: [automations]
      summary: List automations
      parameters:
        - name: workspaceId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Automation"
    post:
      operationId: createAutomation
      tags: [automations]
      summary: Create automation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAutomationBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"
  /automations/{automationId}:
    patch:
      operationId: updateAutomation
      tags: [automations]
      summary: Update automation
      parameters:
        - name: automationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAutomationBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"
    delete:
      operationId: deleteAutomation
      tags: [automations]
      summary: Delete automation
      parameters:
        - name: automationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── NOTIFICATIONS ──────────────────────────────────────────────────────────
...
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary stats
      parameters:
        - name: workspaceId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent activity feed
      parameters:
        - name: workspaceId
          in: query
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
  /dashboard/upcoming-posts:
    get:
      operationId: getUpcomingPosts
      tags: [dashboard]
      summary: Get next scheduled posts
      parameters:
        - name: workspaceId
          in: query
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"

  # ─── AI ─────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ── Auth ──
...
    Post:
      type: object
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        authorId:
          type: integer
        authorName:
          type: string
        caption:
          type: string
        mediaUrls:
          type: array
          items:
            type: string
        mediaType:
          type: string
          enum: [image, video, carousel, text, reel, story]
        platforms:
          type: array
          items:
            type: string
        hashtags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, pending_approval, revision_requested, approved, scheduled, publishing, published, failed, archived]
        scheduledAt:
          type: string
          nullable: true
        publishedAt:
          type: string
          nullable: true
        totalReach:
          type: integer
          nullable: true
        totalEngagement:
          type: integer
          nullable: true
        campaignId:
          type: integer
          nullable: true
        aiGenerated:
          type: boolean
        createdAt:
          type: string
      required: [id, workspaceId, authorId, authorName, caption, mediaUrls, mediaType, platforms, hashtags, status, aiGenerated, createdAt]
...
    Campaign:
      type: object
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        objective:
          type: string
          enum: [awareness, engagement, leads, sales, traffic]
        status:
          type: string
          enum: [draft, active, paused, completed, archived]
        startDate:
          type: string
          nullable: true
        endDate:
          type: string
          nullable: true
        budget:
          type: number
          nullable: true
        currency:
          type: string
          nullable: true
        targetPlatforms:
          type: array
          items:
            type: string
        postsCount:
          type: integer
        leadsCount:
          type: integer
        totalReach:
          type: integer
        totalEngagement:
          type: integer
        roi:
          type: number
          nullable: true
        color:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required: [id, workspaceId, name, objective, status, targetPlatforms, postsCount, leadsCount, totalReach, totalEngagement, tags, createdAt]
    CreateCampaignBody:
      type: object
      properties:
        workspaceId:
          type: integer
        name:
          type: string
        description:
          type: string
        objective:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        budget:
          type: number
        currency:
          type: string
        targetPlatforms:
          type: array
          items:
            type: string
        color:
          type: string
        tags:
          type: array
          items:
            type: string
      required: [workspaceId, name, objective]
    UpdateCampaignBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        budget:
          type: number
        targetPlatforms:
          type: array
          items:
            type: string

    # ── Listening ──
...
    Automation:
      type: object
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
        triggerType:
          type: string
        actionTypes:
          type: array
          items:
            type: string
        runCount:
          type: integer
        lastRunAt:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, workspaceId, name, isActive, triggerType, actionTypes, runCount, createdAt]
    CreateAutomationBody:
      type: object
      properties:
        workspaceId:
          type: integer
        name:
          type: string
        description:
          type: string
        triggerType:
          type: string
        actionTypes:
          type: array
          items:
            type: string
      required: [workspaceId, name, triggerType, actionTypes]
    UpdateAutomationBody:
      type: object
      properties:
        name:
          type: string
        isActive:
          type: boolean
        description:
          type: string

    # ── Notification ──
...
    DashboardSummary:
      type: object
      properties:
        totalFollowers:
          type: integer
        followerGrowth:
          type: number
        scheduledPosts:
          type: integer
        publishedPosts:
          type: integer
        unreadMessages:
          type: integer
        activeCampaigns:
          type: integer
        openLeads:
          type: integer
        mentionsToday:
          type: integer
        engagementRate:
          type: number
        reachThisWeek:
          type: integer
      required: [totalFollowers, followerGrowth, scheduledPosts, publishedPosts, unreadMessages, activeCampaigns, openLeads, mentionsToday, engagementRate, reachThisWeek]
    ActivityItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        title:
          type: string
        description:
          type: string
        actorName:
          type: string
        actorAvatar:
          type: string
          nullable: true
        platform:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, type, title, description, actorName, createdAt]

    # ── AI ──
...